Bug Fixes For State Mahcine, Zero, Commands, Heading
Added Coral Score Drive back in
Added Zero Gyro Back in
Fixed all state machine values so it would be triggers and not only be called once
Reset Gyro to 0 and not 180 (Note this may need to be changed back but will have to check user in put. We want the input of the controller to be positive but most of have rotation correct without april tags)
Turned Smart Drive on by default.